backend/requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Marshmallow==1.2.0
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Caching==2.1.0

# SQLAlchemy and Marshmallow
SQLAlchemy==2.0.23
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Swagger/OpenAPI
flasgger==0.9.7.1
apispec==6.3.0

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter
python-dotenv==1.0.0

# Redis & Caching
redis==5.0.1
hiredis==2.3.2

# Celery for background tasks
celery==5.3.4
kombu==5.3.4

# Security
bcrypt==4.1.2
PyJWT==2.8.0

# Utilities
gunicorn==21.2.0
python-dateutil==2.8.2
python-slugify==8.0.1

# Development & Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1
factory-boy==3.3.0
faker==21.0.0
freezegun==1.2.2
black==23.12.1
flake8==6.1.0
